pyproject.toml
[tool.poetry]
name = "count-clicks"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "count_clicks"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
types-requests = "^2.31.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
